+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Adding additional validation and description for the `loq_legend` argument.
 * Decrease width of error bars in `g_lineplot`.
+* Add option to add rug plot in `g_density_distribution_plot`.
+
 * issues with no news:
 
 # goshawk 0.1.5